dev10 Capstone Project/dev10-Capstone-Project/requirements.txt
bcrypt==4.2.1
fastapi==0.115.6
passlib==1.7.4
PyMySql==1.1.1
pytest==8.3.4
pytest-mock==3.10.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.20
starlette==0.41.3
SQLAlchemy==2.0.19
uvicorn==0.22.0
numpy==1.26.4
pandas==2.0.3
pydantic==1.10.7
scikit-learn==1.3.1
